api: status: try to prevent converting integers to floating point
Since perl can have multiple internal representations for the same
scalar variables, it calculates and caches them on demand.
So when doing a division with a variable, it calculates and caches the
floating point representation of it.

When using JSON::XS (which is the default and what we use), it has to
look at the internal representation and chooses a format, and if
both integer and floating point representations are there it chooses,
floating point.

This is usually not a problem, but for bigger numbers this representation
changes to the format `1.234e+15`.

Since our API designates the `total` and `used` fields as integers, our
Schema2Rust code wants to de-serialize this as an integer again and
fails with such values. (And with this the PDM de-serialization)

To avoid this internal conversion, extract the variables before doing
division with it, so the original object is not touched.

There might be more places where we do such calculations, but it's hard
to search for them, so let's fix it as we find them.
(And comment to avoid accidental refactoring to the broken code)

Example perl code to reproduce the problem:

```
use JSON;
my $foo = 12057594037927909;
print JSON::encode_json($foo);
```

returns `12057594037927909`, but

```
use JSON;
my $foo = 12057594037927909;
my $bar = 123.123 / $foo;
print JSON::encode_json($foo);
```
return `1.20575940379279e+16`.

This was reported in the PDM forum:
https://forum.proxmox.com/threads/177888/

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/all/20251215081435.472283-1-d.csapak@proxmox.com